Login Concept
Created a login concept to enter username/password

This doesn't break down login based upon admin, buyer, seller, instead I was thinking we could just add a drop down when entering a card that would indicate if the card was being bought(iso) or sold. Figured with given time that might be easier and less time consuming.

This is a in-progress concept and isn't finished. Still need to route the page to the index.html page once submission is complete.